pythagorean_triplet: better elimination of inner branches
Feels like we're getting places.

Compare the signs of the pythagorean difference (c^2 - (a^2 + b^2)) of
the top and bottom of the inner loop (the 'b' range), and if they are
the same then there will be no solutions in that range.

Still taking about 59 seconds though, so more to do.

A binary search of the inner loop would cut out a lot of the search
space.